+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed an issue where the click event for buttons inside table cells was not triggered by the spacebar.
+  
 ## 5.15.0 - (March 21, 2024)
 
 * Changed