+++ packages/terra-table/CHANGELOG.md
@@ -7,7 +7,7 @@
 
 * Fixed
   * Fixed an issue where the click event for buttons inside table cells was not triggered by the spacebar.
-
+  
 ## 5.15.0 - (March 21, 2024)
 
 * Changed